chore!(many): moved a bunch of files.
- I now have a new type: "extensions:, i.e: "telescope-extensions" to
  have extensions for each plugin. They are in the lua/user/ directory,
  separated from the config directories.
